requirements.txt
Flask==3.0.3
pytesseract==0.3.10
Pillow==10.4.0
gunicorn==23.0.0
